rename SidebarContents to SidebarContainer and updated examples in TabbedSidebar
