+++ packages/ui/src/lib/sidebar/SidebarContainer.svelte
